// src/payload-types.ts
export interface User {
  id: string;
  username: string;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
// ...
export interface Category {
  id: string;
  title: string;
  slug: string;
  posts?: {
    docs?: (string | Post)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  content?: string | null;
  owner: string | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "posts".
 */
export interface Post {
  id: string;
  title: string;
  slug: string;
  categories?: (string | Category)[] | null;
  content?: string | null;
  owner: string | User;
  updatedAt: string;
  createdAt: string;
}